requirements.txt
nextcord==3.1.0
pyaudio==0.2.14
vosk==0.3.45
numpy>=2.0.0  # Use the newest compatible version
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydantic==1.10.13